Bad retrieve no longer breaks, delete tokens implemented
- Attempting to get tokens by entering a password when no password previously saved used to break system. Solved!
- Users can now delete a token.
